feat: implement tiled image buffer system with core pixel manipulation and blending operations

+++ src/core/cpp/include/image_buffer.h
@@ -54,7 +54,7 @@ public:
                   bool alphaLock = false, bool isEraser = false);
 
   // Draw a filled circle (for brush dabs)
-  void drawCircle(int cx, int cy, float radius, uint8_t r, uint8_t g, uint8_t b,
+  void drawCircle(float cx, float cy, float radius, uint8_t r, uint8_t g, uint8_t b,
                   uint8_t a, float hardness = 1.0f, float grain = 0.0f,
                   bool alphaLock = false, bool isEraser = false,
                   const ImageBuffer *mask = nullptr);
